+++ libcircuit_reliability/src/simulations_handling/spectre_simulations/montecarlo_critical_parameter_value_simulation.cpp
@@ -196,7 +196,7 @@ bool MontecarloCriticalParameterValueSimulation::AnalyzeMontecarloResults(){
 	bool partialResult = true;
 	// create metric magnitudes structure
 	unsigned int upsetsCount = 0;
-	auto metricMagnitudes = golden_magnitudes_structure->GetMetricMagnitudesVector(0);
+	auto metricMagnitudes = golden_magnitudes_structure->GetBasicMetricMagnitudesVector();
 	double maxErrorGlobal [metricMagnitudes->size()];
 	double maxErrorMetric [metricMagnitudes->size()];
 	double minErrorMetric [metricMagnitudes->size()];
@@ -449,7 +449,7 @@ bool MontecarloCriticalParameterValueSimulation::PlotProfileResultsMagnitudes(
 	unsigned int magCount = 0;
 	unsigned int magResultIndex = 2;
 	// for each plotable
-	for( auto const &mag : *(golden_magnitudes_structure->GetPlottableMagnitudesVector(0)) ){
+	for( auto const &mag : *(golden_magnitudes_structure->GetBasicPlottableMagnitudesVector()) ){
 		// title
 		std::string title = mag->get_title_name() +  ", " +  simulation_id +  " montecarlo scatter";
 		// Files