+++ changelog.md
@@ -10,6 +10,7 @@
 # Changelog of version 3.0.0
 * Clean build for *smacd_2016*, including all *v2.9.y* improvements and bug fixes.
 * **CreateFolder** Method is a global template function
+* **NDMagnitudesStructure** now includes constant pointers to useful vectors (plottable and metric magnitudes)
 * Minor *profile scatter* plotting bug has been corrected.
 
 # Version 3.x.y
@@ -19,6 +20,7 @@
 * Changelog of version 3.0.0
 	* Clean build for *smacd_2016*, including all *v2.9.y* improvements and bug fixes.
 	* **CreateFolder** Method is a global template function
+	* **NDMagnitudesStructure** now includes constant pointers to useful vectors (plottable and metric magnitudes)
 	* Minor *profile scatter* plotting bug has been corrected.
 
 # Version 2.x.y